tests/fixtures/vulnerable-python/requirements.txt
requests==2.20.0
pyyaml==5.3.0
django==3.2.0
